+++ src/TERC20Upgradeable.sol
@@ -6,6 +6,9 @@ import "OZUpgradeable/access/AccessControlUpgradeable.sol";
 import "OZUpgradeable/proxy/utils/Initializable.sol";
 import "./lib/TERC20Share.sol";
 
+/**
+* @title TERC20 for an upgradeable deployment, with a compatible proxy (Transparent or Beacon)
+*/
 contract TERC20Upgradeable is
     Initializable,
     ERC20Upgradeable,
@@ -60,12 +63,33 @@ contract TERC20Upgradeable is
      * @notice Returns the number of decimals used to get its user representation.
      * @inheritdoc ERC20Upgradeable
      */
-    function decimals() public view virtual override returns (uint8) {
+    function decimals()
+        public
+        view
+        virtual
+        override(ERC20Upgradeable)
+        returns (uint8)
+    {
         TERC20UpgradeableStorage storage $ = _getTERC20UpgradeableStorage();
         return $._decimals;
     }
 
+    /**
+    * @inheritdoc TERC20Share
+    */
+    function version()
+        public
+        pure
+        override(TERC20Share)
+        returns (string memory)
+    {
+        return TERC20Share.VERSION;
+    }
+
     /* ============ Mint ============ */
+    /**
+    * @inheritdoc TERC20Share
+    */
     function mint(
         address account,
         uint256 value
@@ -74,7 +98,10 @@ contract TERC20Upgradeable is
         emit Mint(msg.sender, account, value);
     }
 
-    function mintBatch(
+    /**
+    * @inheritdoc TERC20Share
+    */
+    function batchMint(
         address[] calldata accounts,
         uint256[] calldata values
     ) public override onlyRole(MINTER_ROLE) {
@@ -89,10 +116,13 @@ contract TERC20Upgradeable is
         for (uint256 i = 0; i < accounts.length; ++i) {
             _mint(accounts[i], values[i]);
         }
-        emit MintBatch(msg.sender, accounts, values);
+        emit BatchMint(msg.sender, accounts, values);
     }
 
     /* ============ Burn ============ */
+    /**
+    * @inheritdoc TERC20Share
+    */
     function burn(
         address account,
         uint256 value
@@ -101,7 +131,10 @@ contract TERC20Upgradeable is
         emit Burn(msg.sender, account, value);
     }
 
-    function burnBatch(
+    /**
+    * @inheritdoc TERC20Share
+    */
+    function batchBurn(
         address[] calldata accounts,
         uint256[] calldata values
     ) public override onlyRole(BURNER_ROLE) {
@@ -116,7 +149,7 @@ contract TERC20Upgradeable is
         for (uint256 i = 0; i < accounts.length; ++i) {
             _burn(accounts[i], values[i]);
         }
-        emit BurnBatch(msg.sender, accounts, values);
+        emit BatchBurn(msg.sender, accounts, values);
     }
 
     /* ============ ACCESS CONTROL ============ */