+++ test/TERC20TestProxy.t.sol
@@ -4,21 +4,27 @@ pragma solidity ^0.8.27;
 import "forge-std/Test.sol";
 import "../src/TERC20Upgradeable.sol";
 import "./TERC20TestShare.sol";
-import { Upgrades } from "openzeppelin-foundry-upgrades/Upgrades.sol";
-
-
+import {Upgrades} from "openzeppelin-foundry-upgrades/Upgrades.sol";
 
 contract TERC20TestProxy is Test, TERC20TestShare {
-
-    function setUp() public{
+    function setUp() public {
         address proxy = Upgrades.deployTransparentProxy(
             "TERC20Upgradeable.sol",
             admin,
-            abi.encodeCall(TERC20Upgradeable.initialize, ( admin, testName, testSymbol, testDecimals))
+            abi.encodeCall(
+                TERC20Upgradeable.initialize,
+                (admin, testName, testSymbol, testDecimals)
+            )
         );
-       token = TERC20Upgradeable(proxy);
+        token = TERC20Upgradeable(proxy);
     }
 
+    /*//////////////////////////////////////////////////////////////
+                        VERSION
+    //////////////////////////////////////////////////////////////*/
+    function testVersion() public view {
+        testShareVersion();
+    }
 
     /*//////////////////////////////////////////////////////////////
                           MINT
@@ -26,13 +32,11 @@ contract TERC20TestProxy is Test, TERC20TestShare {
 
     function testMint() public {
         TERC20TestShare.testShareCanMint();
-        
     }
 
-    function testMintBatch() public {
-        TERC20TestShare.testShareCanMintBatch();
+    function testBatchMint() public {
+        TERC20TestShare.testShareCanBatchMint();
     }
-
 
     /*//////////////////////////////////////////////////////////////
                            BURN
@@ -42,32 +46,31 @@ contract TERC20TestProxy is Test, TERC20TestShare {
         TERC20TestShare.testShareCanBurn();
     }
 
-    function testBurnBatch() public {
-        TERC20TestShare.testShareCanBurnBatch();
+    function testBatchBurn() public {
+        TERC20TestShare.testShareCanBatchBurn();
     }
 
     /*//////////////////////////////////////////////////////////////
                            Access Control
     //////////////////////////////////////////////////////////////*/
 
-    function testAttackerCannotBurnAndBurnBatch() public {
-        TERC20TestShare.testShareAttackerCannotBurnAndBurnBatch();
+    function testAttackerCannotBurnAndBatchBurn() public {
+        TERC20TestShare.testShareAttackerCannotBurnAndBatchBurn();
     }
 
-    function testAttackerCannotMintAndMintBatch() public {
-        TERC20TestShare.testShareAttackerCannotMintAndMintBatch();
+    function testAttackerCannotMintAndBatchMint() public {
+        TERC20TestShare.testShareAttackerCannotMintAndBatchMint();
     }
-
 
     /*//////////////////////////////////////////////////////////////
                           Invalid Parameters
     //////////////////////////////////////////////////////////////*/
-    function testAttackerCannotBurnBatchIfInvalidParameters() public {
-        TERC20TestShare.testShareCannotBurnBatchIfInvalidParameters();
+    function testAttackerCannotBatchBurnIfInvalidParameters() public {
+        TERC20TestShare.testShareCannotBatchBurnIfInvalidParameters();
     }
 
-    function testAttackerCannotMintBatchIfInvalidParameters() public {
-        TERC20TestShare.testShareCannotMintBatchIfInvalidParameters();
+    function testAttackerCannotBatchMintIfInvalidParameters() public {
+        TERC20TestShare.testShareCannotBatchMintIfInvalidParameters();
     }
 
     /*//////////////////////////////////////////////////////////////
@@ -77,11 +80,20 @@ contract TERC20TestProxy is Test, TERC20TestShare {
         address proxy = Upgrades.deployTransparentProxy(
             "TERC20Upgradeable.sol",
             admin,
-            abi.encodeCall(TERC20Upgradeable.initialize, ( admin, testName, testSymbol, testDecimals))
+            abi.encodeCall(
+                TERC20Upgradeable.initialize,
+                (admin, testName, testSymbol, testDecimals)
+            )
         );
         TERC20Upgradeable TERC20 = TERC20Upgradeable(proxy);
         assertEq(TERC20.decimals(), testDecimals);
         assertEq(TERC20.name(), testName);
         assertEq(TERC20.symbol(), testSymbol);
+        assertEq(TERC20.symbol(), testSymbol);
+        assertEq(TERC20.hasRole(DEFAULT_ADMIN_ROLE, admin), true);
+        assertEq(TERC20.hasRole(MINTER_ROLE, admin), true);
+        assertEq(TERC20.hasRole(BURNER_ROLE, admin), true);
+        // Admin has all the roles
+        assertEq(TERC20.hasRole(NOT_ROLE, admin), true);
     }
 }