+++ CHANGELOG.md
@@ -2,6 +2,13 @@
 
 Please follow <https://changelog.md/> conventions.
 
+
+
+## 1.0.0
+
+- Add function `version` to return contract version
+- Rename functions `burnBatch`, `mintBatch`and their corresponding events in `batchBurn`, `batchMint` to use the same name as [ERC-3643](https://eips.ethereum.org/EIPS/eip-3643)
+
 ## 0.1.0
 
 First release !